# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: M15-AUDIT API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/balance:
    get:
      operationId: getBalanceDesComptes
      tags: [reporting]
      summary: "La Balance des Comptes: one row per SYSCOHADA account for the given client/fiscal year, with opening balance, movements and closing balance"
      parameters:
        - name: clientId
          in: query
          required: true
          schema:
            type: integer
        - name: year
          in: query
          required: true
          schema:
            type: integer
        - name: stationId
          in: query
          required: false
          description: "Multi-station (P8): restrict the report to one physical station."
          schema:
            type: integer
      responses:
        "200":
          description: Balance des comptes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceDesComptesResult"
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reports/bilan:
    get:
      operationId: getBilanSimplifie
      tags: [reporting]
      summary: "Le Bilan Simplifié: Actif / Passif aggregated from SYSCOHADA classes 1 to 5 (plus the fiscal year's résultat net) for the given client/fiscal year"
      parameters:
        - name: clientId
          in: query
          required: true
          schema:
            type: integer
        - name: year
          in: query
          required: true
          schema:
            type: integer
        - name: stationId
          in: query
          required: false
          description: "Multi-station (P8): restrict the report to one physical station."
          schema:
            type: integer
      responses:
        "200":
          description: Bilan simplifié
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BilanSimplifieResult"
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reports/compte-resultat:
    get:
      operationId: getCompteDeResultat
      tags: [reporting]
      summary: "Le Compte de Résultat Simplifié: charges (classe 6) vs produits (classe 7) for the given client/fiscal year, netting to the résultat net"
      parameters:
        - name: clientId
          in: query
          required: true
          schema:
            type: integer
        - name: year
          in: query
          required: true
          schema:
            type: integer
        - name: stationId
          in: query
          required: false
          description: "Multi-station (P8): restrict the report to one physical station."
          schema:
            type: integer
      responses:
        "200":
          description: Compte de résultat simplifié
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompteDeResultatResult"
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reports/grand-livre:
    get:
      operationId: getGrandLivre
      tags: [reporting]
      summary: "Le Grand Livre: every validated SYSCOHADA account for the given client/fiscal year, grouped with its chronological movements and running balance"
      parameters:
        - name: clientId
          in: query
          required: true
          schema:
            type: integer
        - name: year
          in: query
          required: true
          schema:
            type: integer
        - name: stationId
          in: query
          required: false
          description: "Multi-station (P8): restrict the report to one physical station."
          schema:
            type: integer
      responses:
        "200":
          description: Grand livre
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GrandLivreResult"
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reports/pilotage:
    get:
      operationId: getPilotageDashboard
      tags: [reporting]
      summary: "Module P4 (Pilotage Dirigeant): plain-language dashboard aggregates for the PME director -- net cash position, monthly revenue trend and top expense categories"
      parameters:
        - name: clientId
          in: query
          required: true
          schema:
            type: integer
        - name: year
          in: query
          required: true
          schema:
            type: integer
        - name: basis
          in: query
          required: false
          description: "Module M21: 'engagement' (comptabilité d'engagement, default) or 'tresorerie' (comptabilité de trésorerie -- credit operations count only once settled)"
          schema:
            type: string
            enum: [engagement, tresorerie]
        - name: month
          in: query
          required: false
          description: "Filters the KPI cards (chiffre d'affaires, marge brute, trésorerie) to a specific month (1-12) within the selected year, compared against the month right before it. Omit to default to the most recent month with activity."
          schema:
            type: integer
            minimum: 1
            maximum: 12
        - name: stationId
          in: query
          required: false
          description: "Multi-station (P8): restrict the dashboard to one physical station."
          schema:
            type: integer
      responses:
        "200":
          description: Pilotage dashboard aggregates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PilotageDashboardResult"
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reports/export-liasse:
    post:
      operationId: exportLiasseFiscale
      tags: [reporting]
      summary: "Records an M9 audit trail entry for a liasse fiscale (PDF) export request. Module M3 reporting export is a mocked action in this MVP -- no PDF is generated -- but every export attempt must still be auditable."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportLiasseFiscaleInput"
      responses:
        "200":
          description: Export request logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportLiasseFiscaleResult"
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    BalanceSide:
      type: string
      enum: [debiteur, crediteur]
    BalanceRow:
      type: object
      required:
        [
          accountNumber,
          accountName,
          accountClass,
          initialBalance,
          totalDebit,
          totalCredit,
          finalBalance,
          finalBalanceSide,
        ]
      properties:
        accountNumber: { type: string }
        accountName: { type: string }
        accountClass: { type: integer }
        initialBalance: { type: integer }
        totalDebit: { type: integer }
        totalCredit: { type: integer }
        finalBalance: { type: integer }
        finalBalanceSide:
          $ref: "#/components/schemas/BalanceSide"
    BalanceDesComptesResult:
      type: object
      required: [clientId, year, rows]
      properties:
        clientId: { type: integer }
        year: { type: integer }
        rows:
          type: array
          items:
            $ref: "#/components/schemas/BalanceRow"
    BilanLine:
      type: object
      required: [key, label, amount]
      properties:
        key: { type: string }
        label: { type: string }
        amount: { type: integer }
    BilanSimplifieResult:
      type: object
      required: [clientId, year, actif, passif, totalActif, totalPassif]
      properties:
        clientId: { type: integer }
        year: { type: integer }
        actif:
          type: array
          items:
            $ref: "#/components/schemas/BilanLine"
        passif:
          type: array
          items:
            $ref: "#/components/schemas/BilanLine"
        totalActif: { type: integer }
        totalPassif: { type: integer }
    CompteResultatLine:
      type: object
      required: [accountNumber, label, amount]
      properties:
        accountNumber: { type: string }
        label: { type: string }
        amount: { type: integer }
    CompteDeResultatResult:
      type: object
      required: [clientId, year, charges, produits, totalCharges, totalProduits, resultatNet]
      properties:
        clientId: { type: integer }
        year: { type: integer }
        charges:
          type: array
          items:
            $ref: "#/components/schemas/CompteResultatLine"
        produits:
          type: array
          items:
            $ref: "#/components/schemas/CompteResultatLine"
        totalCharges: { type: integer }
        totalProduits: { type: integer }
        resultatNet: { type: integer }
    GrandLivreMovement:
      type: object
      required: [date, label, debitAmount, creditAmount, runningBalance, runningBalanceSide]
      properties:
        date: { type: string, format: date-time }
        label: { type: string }
        debitAmount: { type: integer }
        creditAmount: { type: integer }
        runningBalance: { type: integer }
        runningBalanceSide:
          $ref: "#/components/schemas/BalanceSide"
    GrandLivreAccount:
      type: object
      required:
        [
          accountNumber,
          accountName,
          accountClass,
          initialBalance,
          initialBalanceSide,
          movements,
          totalDebit,
          totalCredit,
          finalBalance,
          finalBalanceSide,
        ]
      properties:
        accountNumber: { type: string }
        accountName: { type: string }
        accountClass: { type: integer }
        initialBalance: { type: integer }
        initialBalanceSide:
          $ref: "#/components/schemas/BalanceSide"
        movements:
          type: array
          items:
            $ref: "#/components/schemas/GrandLivreMovement"
        totalDebit: { type: integer }
        totalCredit: { type: integer }
        finalBalance: { type: integer }
        finalBalanceSide:
          $ref: "#/components/schemas/BalanceSide"
    GrandLivreResult:
      type: object
      required: [clientId, year, accounts]
      properties:
        clientId: { type: integer }
        year: { type: integer }
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/GrandLivreAccount"
    MonthlyRevenuePoint:
      type: object
      required: [year, month, label, total]
      properties:
        year: { type: integer }
        month: { type: integer, minimum: 1, maximum: 12 }
        label: { type: string }
        total: { type: integer }
    ExpenseBreakdownEntry:
      type: object
      required: [categoryKey, label, total]
      properties:
        categoryKey: { type: string }
        label: { type: string }
        total: { type: integer }
    PilotageDashboardResult:
      type: object
      required:
        [
          clientId,
          year,
          basis,
          tresorerieNette,
          chiffreAffairesParMois,
          topDepenses,
          chargesParMois,
          margeBruteParMois,
          tresorerieParMois,
          depensesParNature,
          seuilRentabilite,
          kpis,
        ]
      properties:
        clientId: { type: integer }
        year: { type: integer }
        basis:
          type: string
          enum: [engagement, tresorerie]
        tresorerieNette: { type: integer }
        chiffreAffairesParMois:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyRevenuePoint"
        topDepenses:
          type: array
          items:
            $ref: "#/components/schemas/ExpenseBreakdownEntry"
        chargesParMois:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyRevenuePoint"
        margeBruteParMois:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyMarginPoint"
        tresorerieParMois:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyRevenuePoint"
        depensesParNature:
          type: array
          items:
            $ref: "#/components/schemas/NatureBreakdownEntry"
        seuilRentabilite:
          $ref: "#/components/schemas/SeuilRentabilite"
        kpis:
          $ref: "#/components/schemas/ExecutiveDashboardKpis"
    MonthlyMarginPoint:
      type: object
      required: [year, month, label, chiffreAffaires, margeBrute, tauxMarge]
      properties:
        year: { type: integer }
        month: { type: integer, minimum: 1, maximum: 12 }
        label: { type: string }
        chiffreAffaires: { type: integer }
        margeBrute: { type: integer }
        tauxMarge:
          type: number
          nullable: true
    NatureBreakdownEntry:
      type: object
      required: [natureKey, label, total]
      properties:
        natureKey: { type: string }
        label: { type: string }
        total: { type: integer }
    SeuilRentabilite:
      type: object
      required:
        [
          chiffreAffairesAnnuel,
          chargesFixesAnnuelles,
          chargesVariablesAnnuelles,
          tauxMargeSurCoutsVariables,
          seuilRentabilite,
        ]
      properties:
        chiffreAffairesAnnuel: { type: integer }
        chargesFixesAnnuelles: { type: integer }
        chargesVariablesAnnuelles: { type: integer }
        tauxMargeSurCoutsVariables:
          type: number
          nullable: true
        seuilRentabilite:
          type: number
          nullable: true
    DashboardKpi:
      type: object
      required: [moisCourant, moisPrecedent, variationPct]
      properties:
        moisCourant: { type: number }
        moisPrecedent: { type: number }
        variationPct:
          type: number
          nullable: true
    MargeBruteKpi:
      allOf:
        - $ref: "#/components/schemas/DashboardKpi"
        - type: object
          required: [tauxMargeMoisCourant, tauxMargeMoisPrecedent]
          properties:
            tauxMargeMoisCourant:
              type: number
              nullable: true
            tauxMargeMoisPrecedent:
              type: number
              nullable: true
    TresorerieKpi:
      allOf:
        - $ref: "#/components/schemas/DashboardKpi"
        - type: object
          required: [seuilCritique, enAlerte]
          properties:
            seuilCritique: { type: number }
            enAlerte: { type: boolean }
    ExecutiveDashboardKpis:
      type: object
      required: [chiffreAffaires, margeBrute, tresorerie]
      properties:
        chiffreAffaires:
          $ref: "#/components/schemas/DashboardKpi"
        margeBrute:
          $ref: "#/components/schemas/MargeBruteKpi"
        tresorerie:
          $ref: "#/components/schemas/TresorerieKpi"
    ExportLiasseFiscaleInput:
      type: object
      required: [clientId, year, reportType]
      properties:
        clientId: { type: integer }
        year: { type: integer }
        reportType:
          type: string
          enum: [balance, bilan, compte_resultat]
    ExportLiasseFiscaleResult:
      type: object
      required: [logged]
      properties:
        logged: { type: boolean }

    # -------------------------------------------------------------------------
    # Module M17 — Fixed Assets & Depreciation Engine
    # -------------------------------------------------------------------------
...